Various fixes to Random.hpp
* Removes a compiler warning since we were mixing 32 and 64 bit types
* Removed broken and useless randN24-function and made rand0X to work
  with all integral and floating point types
* Added rand64
* Fixed rand0X64 to be uniform

+++ multitude/Nimble/Random.cpp
@@ -15,12 +15,12 @@
 namespace Nimble {
   RandomUniform  RandomUniform::m_instance;
 
-  unsigned long RandomUniform::randomSeed()
+  uint64_t RandomUniform::randomSeed()
   {
     static std::random_device rnd;
     auto now = std::chrono::high_resolution_clock::now();
     auto ns = std::chrono::duration_cast<std::chrono::nanoseconds>(now.time_since_epoch()).count();
-    return rnd() ^ (static_cast<unsigned long>(rnd()) << 32ul) ^ ns;
+    return rnd() ^ (static_cast<uint64_t>(rnd()) << uint64_t(32)) ^ ns;
   }
 
 }
